SL-SL-1: moves session search related suff to one folder
